+++ notification-alert-job/src/main/java/by/gdev/alert/job/notification/model/MessageData.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class MessageData {
-	private int chat_id;
+	private Long chat_id;
 	private String text;
 }
